bridge/dev_requirements.txt
pytest==9.0.2
pytest-aiohttp==1.1.0
coverage>=6.3
